pyproject.toml
[project]
name = "calorie-counter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "audioread==3.0.1",
    "backports-tarfile==1.2.0",
    "blis==1.3.0",
    "build==1.2.2.post1",
    "cachecontrol==0.14.3",
    "catalogue==2.0.10",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "cleo==2.1.0",
    "click==8.2.1",
    "cloudpathlib==0.21.1",
    "comm==0.2.2",
    "confection==0.1.5",
    "crashtest==0.4.1",
    "cryptography==45.0.4",
    "cymem==2.0.11",
    "dawg2-python==0.9.0",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "distlib==0.3.9",
    "dulwich==0.22.8",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "findpython==0.6.3",
    "fsspec==2025.5.1",
    "gensim==4.3.2",
    "h11==0.16.0",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.33.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "installer==0.7.0",
    "ipykernel==6.29.5",
    "ipython==8.37.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.2.1",
    "jedi==0.19.2",
    "jeepney==0.9.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "keyring==25.6.0",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "llvmlite==0.44.0",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "murmurhash==1.0.13",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numba==0.61.2",
    "numpy==2.2.6",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "opencv-python==4.11.0.86",
    "opencv-python-headless==4.11.0.86",
    "packaging==25.0",
    "pandas==2.3.0",
    "parso==0.8.4",
    "pbs-installer==2025.6.12",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "pkginfo==1.12.1.2",
    "platformdirs==4.3.8",
    "poetry==2.1.3",
    "poetry-core==2.1.3",
    "pooch==1.8.2",
    "preshed==3.0.10",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pymorphy3==2.0.4",
    "pymorphy3-dicts-ru==2.4.417150.4580142",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "rapidfuzz==3.13.0",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "secretstorage==3.3.3",
    "setuptools==70.3.0",
    "shellingham==1.5.4",
    "simsimd==6.4.9",
    "six==1.17.0",
    "smart-open==7.1.0",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soxr==0.5.0.post1",
    "spacy==3.8.7",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.1",
    "stack-data==0.6.3",
    "stringzilla==3.12.5",
    "sympy==1.14.0",
    "thinc==8.3.6",
    "threadpoolctl==3.6.0",
    "timm==1.0.15",
    "tokenizers==0.21.1",
    "tomli==2.2.1",
    "tomlkit==0.13.3",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchmetrics==1.7.4",
    "torchvision==0.22.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.52.4",
    "trove-classifiers==2025.5.9.12",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "virtualenv==20.31.2",
    "wasabi==1.1.3",
    "wcwidth==0.2.13",
    "weasel==0.4.1",
    "wrapt==1.17.2",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]
